+++ arkenv/src/test/kotlin/com/apurebase/arkenv/ParserTest.kt
@@ -1,26 +1,26 @@
 package com.apurebase.arkenv
 
+import com.apurebase.arkenv.argument.ArkenvArgument
 import com.apurebase.arkenv.feature.ArkenvFeature
 import com.apurebase.arkenv.test.expectThat
 import com.apurebase.arkenv.test.parse
+import com.apurebase.arkenv.util.argument
 import org.junit.jupiter.api.Test
 import strikt.assertions.isEqualTo
 
 class ParserTest {
 
     private val customFeature = object : ArkenvFeature {
-        override fun onParse(arkenv: Arkenv, delegate: ArgumentDelegate<*>): String? = "9000"
+        override fun onParse(arkenv: Arkenv, delegate: ArkenvArgument<*>): String? = "9000"
     }
 
     @Test fun `custom parser`() {
-        val ark = object : Arkenv("Test", configureArkenv {
-            install(customFeature)
-        }) {
+        val ark = object : Arkenv("Test", configureArkenv { +customFeature }) {
             val port: Int by argument()
         }
 
         ark.parse().expectThat {
-            get { port }.isEqualTo(9000)
+            get { port } isEqualTo 9000
         }
     }