+++ arkenv/src/test/kotlin/com/apurebase/arkenv/feature/FileBasedTests.kt
@@ -2,7 +2,8 @@ package com.apurebase.arkenv.feature
 
 import com.apurebase.arkenv.Arkenv
 import com.apurebase.arkenv.ArkenvBuilder
-import com.apurebase.arkenv.argument
+import com.apurebase.arkenv.util.argument
+import com.apurebase.arkenv.test.Expected
 import com.apurebase.arkenv.test.expectThat
 import com.apurebase.arkenv.test.parse
 import strikt.assertions.isEqualTo
@@ -22,14 +23,14 @@ interface FileBasedTests {
     fun verify(
         path: String,
         port: Int = defaultPort,
-        pw: String = "this_is_expected",
+        pw: String = Expected.mysqlPassword,
         locations: List<String> = listOf(),
         vararg args: String = arrayOf()
     ) = PropertiesArk(configure(path, locations)).parse(*args).apply {
         expectThat {
-            get { this.mysqlPassword }.isEqualTo(pw)
-            get { this.port }.isEqualTo(port)
-            get { this.multiLine.trim() }.isEqualTo("this stretches lines")
+            get { mysqlPassword } isEqualTo pw
+            get { this.port } isEqualTo port
+            get { multiLine.trim() } isEqualTo "this stretches lines"
         }
     }
 }